+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Following Semantic Versioning 2.
 - Surveys in Challenges: show survey's answers and exportation.
 
 ## Version 0.0.7 (MINOR)
+- Fix: show solutions only when everything is published
 - Truncate description in cards to 100 chars max
 
 ## Version 0.0.6 (MINOR)